+++ src/types/input/electricity.ts
@@ -2,5 +2,4 @@ export type DETElectricityInput = {
   electricityType?: string | null;
   electricityUnitRate?: number | null;
   userElectricityConsumption?: number | null;
-  preRenovationElectricityOffset?: number | null;
 };